+++ base/components/account/AccountHeader.vue
@@ -10,7 +10,7 @@ const localePath = useLocalePath();
 
 <template>
   <h1 class="text-2xl font-semibold">My Account</h1>
-  <ULink :to="localePath('/account')" class="mt-2">
+  <ULink :to="localePath('/account')" class="mt-2" aria-label="Account">
     {{ activeCustomer?.emailAddress }}
   </ULink>
 </template>