+++ nuxt.config.ts
@@ -1,10 +1,10 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
 import tailwindcss from "@tailwindcss/vite";
+import { shopIdentity } from "./schema/identity";
 
-// https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
   compatibilityDate: "2024-11-01",
-  devtools: { enabled: true },
+  devtools: { enabled: import.meta.env.DEV },
 
   future: {
     compatibilityVersion: 4,
@@ -21,7 +21,34 @@ export default defineNuxtConfig({
     storesDirs: ["./base/stores/**"],
   },
 
-  image: {
-    domains: ["localhost:3001"],
+  // NuxtSEO Modules Config
+  robots: {
+    disallow: ["/account", "/checkout", "/confirmation", "/cart", "/search"],
+    allow: "/",
+    blockNonSeoBots: true,
+  },
+
+  schemaOrg: {
+    identity: shopIdentity,
+  },
+
+  site: {
+    url: process.env.NUXT_PUBLIC_I18N_BASE_URL,
+    name: process.env.NUXT_PUBLIC_SITE_NAME,
+    description: "Nuxt Level Headless E-commerce",
+    env: process.env.NODE_ENV,
+    indexable: process.env.NODE_ENV === "production",
+    trailingSlash: false,
+  },
+
+  sitemap: {
+    sources: [
+      [
+        `${process.env.NUXT_PUBLIC_API_BASE}/vendure/sitemap?base_url=${process.env.NUXT_PUBLIC_I18N_BASE_URL}`,
+        {
+          headers: { "vendure-token": process.env.NUXT_PUBLIC_CHANNEL_TOKEN! },
+        },
+      ],
+    ],
   },
 });